+++ backend/canisters/registry/impl/src/updates/update_token.rs
@@ -13,16 +13,16 @@ fn update_token(args: Args) -> Response {
 fn update_token_impl(args: Args, state: &mut RuntimeState) -> Response {
     if let Some(token) = state.data.tokens.get_mut(args.ledger_canister_id) {
         if let Some(info_url) = args.info_url {
-            token.info_url = Some(info_url);
+            token.info_url = info_url;
         }
         if let Some(how_to_buy_url) = args.how_to_buy_url {
-            token.how_to_buy_url = Some(how_to_buy_url);
+            token.how_to_buy_url = how_to_buy_url;
         }
         if let Some(transaction_url_format) = args.transaction_url_format {
-            token.transaction_url_format = Some(transaction_url_format);
+            token.transaction_url_format = transaction_url_format;
         }
         if let Some(logo) = args.logo {
-            token.logo = Some(logo);
+            token.logo = logo;
         }
         token.last_updated = state.env.now();
         Success