+++ frontend/openchat-agent/src/services/registry/candid/idl.js
@@ -6,9 +6,9 @@ export const idlFactory = ({ IDL }) => {
     'fee' : IDL.Nat,
     'decimals' : IDL.Nat8,
     'added' : TimestampMillis,
-    'how_to_buy_url' : IDL.Opt(IDL.Text),
-    'info_url' : IDL.Opt(IDL.Text),
-    'logo' : IDL.Opt(IDL.Text),
+    'how_to_buy_url' : IDL.Text,
+    'info_url' : IDL.Text,
+    'logo' : IDL.Text,
     'name' : IDL.Text,
     'last_updated' : TimestampMillis,
     'nervous_system' : IDL.Opt(
@@ -20,7 +20,7 @@ export const idlFactory = ({ IDL }) => {
     ),
     'ledger_canister_id' : CanisterId,
     'symbol' : IDL.Text,
-    'transaction_url_format' : IDL.Opt(IDL.Text),
+    'transaction_url_format' : IDL.Text,
   });
   const UpdatesResponse = IDL.Variant({
     'Success' : IDL.Record({